feat: implement filename lookup in database (#3)
As suggested in the feature request, this adds a `db` subcommand to the
`mpq` command. It manages a small SQLite database, and can import
filenames from multiple sources, and stores the matching hashnames.

When listing archives without a listfile, the database can then be used
to lookup filesnames by hash.